[Refactor]: AVCatureSession을 지우고 아이트랙킹 기능만 남겨서 MVC로 정리
